refactor(pulltorefresh): folder structure
