+++ packages/pulltorefresh/src/index.ts
@@ -1,1 +1,2 @@
-export { default as PullToRefresh } from './PullToRefresh'
+export { default as PullToRefresh } from './components/PullToRefresh'
+export { ICustomSpinner, ICustomSpinnerProps } from './types'
