+++ packages/pulltorefresh/src/types.ts
@@ -1,0 +1,8 @@
+import React from 'react'
+
+export interface ICustomSpinnerProps {
+  t: number
+  refreshing: boolean
+}
+
+export type ICustomSpinner = React.ComponentType<ICustomSpinnerProps>
